pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "supervisor"
version = "0.1.0"
description = "A unified service manager for Python/FastAPI projects with process supervision, resource monitoring, auto-fix capabilities, and Caddy integration"
readme = "README.md"
license = "CC0-1.0"
authors = [
    {name = "Philip Orange", email = "git@philiporange.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "peewee>=3.17.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "psutil>=5.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["supervisor*"]

[tool.setuptools.package-data]
supervisor = ["templates/*", "static/*", "prompts/*"]
